refactor: aggregateAll→buildTalliesリネームとテスト分離
- aggregateAll.tsをbuildTallies.tsにリネームし、aggregate/toTallyを非公開化
- aggregate.tsを削除しbuildTallies内部のaggregate関数に統合
- aggregate.test.tsをaggregateGt.test.tsとaggregateCross.test.tsに分離
- export.test.tsをbuildTallies経由に変更
- bench/run.tsをaggregateGt/aggregateCross直接呼び出しに変更
