+++ tests/export.test.ts
@@ -1,7 +1,6 @@
 import { describe, it, expect, beforeAll, afterAll } from "vitest";
-import { aggregate } from "../src/lib/agg/aggregate";
+import { buildTallies } from "../src/lib/agg/buildTallies";
 import type { Question, Tally } from "../src/lib/agg/types";
-import { toTally } from "../src/lib/agg/aggregateAll";
 import { setupDuckDB, teardownDuckDB } from "./helpers/duckdb";
 import { buildExportGrids, type ExportGrid } from "../src/lib/export/exportGrid";
 import { formatCSV } from "../src/lib/export/formatters/csv";
@@ -44,19 +43,8 @@ let crossTallies: Tally[];
 beforeAll(async () => {
   conn = await setupDuckDB();
 
-  // GT tallies
-  gtTallies = [];
-  const q1GtResult = await aggregate(conn, q1, null, "");
-  gtTallies.push(toTally(q1, q1GtResult));
-  const q3GtResult = await aggregate(conn, q3, null, "");
-  gtTallies.push(toTally(q3, q3GtResult));
-
-  // Cross tallies: q2 × q1
-  crossTallies = [];
-  const q2GtResult = await aggregate(conn, q2, null, "");
-  crossTallies.push(toTally(q2, q2GtResult));
-  const q2CrossResult = await aggregate(conn, q2, q1, "");
-  crossTallies.push(toTally(q2, q2CrossResult, q1));
+  gtTallies = await buildTallies(conn, [q1, q3], [], "");
+  crossTallies = await buildTallies(conn, [q2], [q1], "");
 }, 30_000);
 
 afterAll(async () => {
